Updated packages.sh script for better checking the installation of packages. It was tested and seems to be working.

+++ scripts/packages.sh
@@ -1,3 +1,5 @@
+echo "EXECUTING SCRIPT : packages.sh"
+
 #fix apt slow mirrors
 if [ "" == "$(grep edatel /etc/apt/sources.list | grep -v grep)" ]; then
     mv /etc/apt/sources.list /etc/apt/sources.list.old
@@ -16,16 +18,26 @@ fi
 #echo 'Dpkg::Progress-Fancy "1";' >> /etc/apt/apt.conf
 #fi
 
+echo "Fixing missing"
 apt-get update --fix-missing -y &> /dev/null 
+echo "Upgrading packages"
+#apt-get -y upgrade
+
 # NOTE: for lmgc90, python-numpy should be used (version 1.8.x) since lmgc90 does not support higher versions
-PACKAGES=(git python-pip build-essential m4 gfortran libncurses5-dev make cmake cmake-curses-gui libvtk5-dev libvtk6-dev libeigen2-dev libeigen3-dev libopenmpi-dev g++ git-core openmpi-common openmpi-bin libopenmpi-dev mercurial htop paraview fftw-dev python-dev libpython-dev  python-numpy python-scipy python-matplotlib libatlas-dev libblas-dev liblapack-dev tcl-vtk python-vtk swig doxygen python-sphinx subversion libmetis5 libmetis5-dbg libscotch-dev wget patch libgsl0-dev libsuitesparse-dev libboost-python-dev python-tk libmumps-dev libparmetis-dev libgtk-3-dev libfltk1.3-dev libxml2-dev libcgal-dev libhdf5-serial-dev libhdf5-dev libcgal-dev libboost-dev libxres-dev ipython-notebook ntp chromium-browser ispell aspell sox emacs eterm rxvt tmux)
+PACKAGES=(git python-pip build-essential m4 gfortran libncurses5-dev libigraph0-dev make cmake cmake-curses-gui libvtk6-dev libeigen2-dev libeigen3-dev libopenmpi-dev g++ git-core openmpi-common openmpi-bin libopenmpi-dev mercurial htop paraview fftw-dev python-dev libpython-dev  python-numpy python-scipy python-matplotlib libatlas-dev libblas-dev liblapack-dev tcl-vtk python-vtk swig doxygen python-sphinx subversion libmetis5 libmetis5-dbg libscotch-dev wget patch libgsl0-dev libsuitesparse-dev libboost-python-dev python-tk libmumps-dev libparmetis-dev libgtk-3-dev libfltk1.3-dev libxml2-dev libcgal-dev libhdf5-serial-dev libhdf5-dev libcgal-dev libboost-dev libxres-dev ipython-notebook ntp chromium-browser ispell aspell sox emacs eterm rxvt tmux)
 # texlive)
 echo "Installing the following packages : ${PACKAGES[@]} ..."
 for a in ${PACKAGES[@]}; do
-    printf "\nINSTALLING : $a"
-    apt-get install -y $a &>> /var/log/log-install-packs
-    printf "STATUS: $?\n\n"
+    if [ ! $(dpkg -l $a | tail -n 1 | awk '{print ($1 == "ii")}') ]; then 
+	printf "\nINSTALLING : $a"
+	apt-get install -y $a &>> /var/log/log-install-packs
+	printf "Installation STATUS: $?\n\n"
+    fi
 done
+
+echo "Removing not needed packages "
+apt-get -y autoremove
+
 
 # Fix some errors with dictionaries 
 ##apt-get install -y  ${PACKAGES[@]} > /dev/null
